Clean some unused code
