+++ internal/ext_protocols/socks5/errors.go
@@ -1,0 +1,5 @@
+package socks5
+
+import "fmt"
+
+var unknownAddressType = fmt.Errorf("unknown address type")
