+++ internal/ext_protocols/socks5/models.go
@@ -1,76 +1,9 @@
 package socks5
 
-import (
-	"bytes"
-	"encoding/gob"
-	"fmt"
-	"io"
-	"net"
-)
-
-// AuthContext is a Request encapsulates authentication state provided during negotiation
-type AuthContext struct {
-	// Provided auth method
-	Method uint8
-	// Payload provided during negotiation.
-	// Keys depend on the used auth method.
-	// For UserPassauth contains Username
-	Payload map[string]string
-}
-
-// AddrSpec is used to return the target AddrSpec
-// which may be specified as IPv4, IPv6, or a FQDN
-type AddrSpec struct {
-	FQDN string
-	IP   net.IP
-	Port int
-}
-
-// A Request represents request received by a server
-type Request struct {
-	Version      uint8
-	Command      uint8
-	AuthContext  *AuthContext
-	RemoteAddr   *AddrSpec
-	DestAddr     *AddrSpec
-	realDestAddr *AddrSpec
-	bufConn      io.Reader
-}
-
-func (r *Request) toBytes() []byte {
-	b := bytes.Buffer{}
-	e := gob.NewEncoder(&b)
-	err := e.Encode(r)
-	if err != nil {
-		fmt.Println(`failed gob Encode`, err)
-	}
-	return b.Bytes()
-}
-
-func GetRequestFromPack(dataPack []byte) Request {
-	r := Request{}
-	b := bytes.Buffer{}
-	b.Write(dataPack)
-	d := gob.NewDecoder(&b)
-	err := d.Decode(&r)
-	if err != nil {
-		fmt.Println(`failed gob Decode`, err)
-	}
-	return r
-}
-
-// Some data structure used in Apron, may be changed during development
-
-type MsgContentType uint32
-
-const (
-	InitRequest MsgContentType = iota // Indicates the message in wrapper is ApronSocks5ConnectRequest
-)
-
 // ApronSocks5ConnectRequest saves some header info of request which will be sent to Apron network at first
 type ApronSocks5ConnectRequest struct {
-	Version uint8
-	Command uint8
-	DestAddr string	// The dest address may be IPv4 address, IPv6 address or domain name
+	Version  uint8
+	Command  uint8
+	DestAddr string // The dest address may be IPv4 address, IPv6 address or domain name
 	DestPort int
 }
